+++ src/devices/printer.rs
@@ -1,6 +1,15 @@
-use std::fmt::{self, Display, Formatter};
+use std::{
+    fmt::{self, Display, Formatter}
+};
 
-use crate::{serial::SerialDevice, warnln};
+use crate::{ppu::PaletteAlpha, serial::SerialDevice, warnln};
+
+const PRINTER_PALETTE: PaletteAlpha = [
+    [0xff, 0xff, 0xff, 0xff],
+    [0xaa, 0xaa, 0xaa, 0xff],
+    [0x55, 0x55, 0x55, 0xff],
+    [0x00, 0x00, 0x00, 0xff],
+];
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 enum PrinterState {
@@ -63,7 +72,7 @@ impl Display for PrinterState {
 #[derive(Clone, Copy, PartialEq, Eq)]
 enum PrinterCommand {
     Init = 0x01,
-    Start = 0x02,
+    Print = 0x02,
     Data = 0x04,
     Status = 0x0f,
     Other = 0xff,
@@ -73,7 +82,7 @@ impl PrinterCommand {
     pub fn description(&self) -> &'static str {
         match self {
             PrinterCommand::Init => "Init",
-            PrinterCommand::Start => "Start",
+            PrinterCommand::Print => "Print",
             PrinterCommand::Data => "Data",
             PrinterCommand::Status => "Status",
             PrinterCommand::Other => "Other",
@@ -83,7 +92,7 @@ impl PrinterCommand {
     fn from_u8(value: u8) -> Self {
         match value {
             0x01 => PrinterCommand::Init,
-            0x02 => PrinterCommand::Start,
+            0x02 => PrinterCommand::Print,
             0x04 => PrinterCommand::Data,
             0x0f => PrinterCommand::Status,
             _ => PrinterCommand::Other,
@@ -107,6 +116,10 @@ pub struct PrinterDevice {
     status: u8,
     byte_out: u8,
     data: [u8; 0x280],
+    image: [u8; 160 * 200],
+    image_offset: u16,
+    image_buffer: Vec<u8>,
+    callback: fn(image_buffer: &Vec<u8>)
 }
 
 impl PrinterDevice {
@@ -121,6 +134,10 @@ impl PrinterDevice {
             status: 0x0,
             byte_out: 0x0,
             data: [0x00; 0x280],
+            image: [0x00; 160 * 200],
+            image_offset: 0,
+            image_buffer: Vec::new(),
+            callback: |_| {}
         }
     }
 
@@ -133,7 +150,27 @@ impl PrinterDevice {
         self.checksum = 0x0;
         self.status = 0x0;
         self.byte_out = 0x0;
-        self.data = [0x00; 0x280]
+        self.data = [0x00; 0x280];
+        self.image = [0x00; 160 * 200];
+        self.image_offset = 0;
+        
+        self.clear_image_buffer()
+    }
+
+    pub fn set_callback(&mut self, callback: fn(image_buffer: &Vec<u8>)) {
+        self.callback = callback;
+    }
+
+    pub fn image_buffer(&self) -> &Vec<u8> {
+        &self.image_buffer
+    }
+
+    pub fn image_buffer_mut(&mut self) -> &mut Vec<u8> {
+        &mut self.image_buffer
+    }
+
+    pub fn clear_image_buffer(&mut self) {
+        self.image_buffer.clear();
     }
 
     fn run_command(&mut self, command: PrinterCommand) {
@@ -141,14 +178,30 @@ impl PrinterDevice {
             PrinterCommand::Init => {
                 self.status = 0x00;
                 self.byte_out = self.status;
+                self.image_offset = 0;
+                self.image_buffer.clear();
             }
-            PrinterCommand::Start => {
+            PrinterCommand::Print => {
+                let palette_index = self.data[2];
+
+                for index in 0..self.image_offset {
+                    let value = self.image[index as usize];
+                    let pixel_offset = (palette_index >> (value << 1)) & 0x03;
+                    let pixel = PRINTER_PALETTE[pixel_offset as usize];
+                    self.image_buffer.push(pixel[0]);
+                    self.image_buffer.push(pixel[1]);
+                    self.image_buffer.push(pixel[2]);
+                    self.image_buffer.push(pixel[3]);
+                }
+
+                (self.callback)(&self.image_buffer);
+
                 self.byte_out = self.status;
                 self.status = 0x06;
             }
             PrinterCommand::Data => {
                 if self.command_length == 0x280 {
-                    println!("Printer: Going to copy the image for printing");
+                    self.flush_image();
                 }
                 // in case the command is of size 0 we assume this is
                 // an EOF and we ignore this data operation
@@ -166,7 +219,8 @@ impl PrinterDevice {
                 self.byte_out = self.status;
 
                 // in case the current status is printing let's
-                // mark it as done
+                // mark it as done, resetting the status back to
+                // the original value
                 if self.status == 0x06 {
                     // @TODO: check if this value should be 0x04 instead
                     // this seems to be a bug with the print demo
@@ -176,6 +230,34 @@ impl PrinterDevice {
             PrinterCommand::Other => {
                 warnln!("Printer: Invalid command: {:02x}", self.state as u8);
             }
+        }
+    }
+
+    fn flush_image(&mut self) {
+        // sets the initial value of the index that will point to
+        // the data that is going to be copied to the image buffer
+        let mut index = 0;
+
+        // iterates over the two rows that are going to be printed
+        // keep in mind that the printer only allows 2 lines at each
+        // time to be printed
+        for _ in 0..2 {
+            for col in 0..20 {
+                for y in 0..8 {
+                    let mut first = self.data[index];
+                    let mut second = self.data[index + 1];
+                    for x in 0..8 {
+                        let offset = self.image_offset as usize + (col * 8) + (y * 160) + x;
+                        self.image[offset] = (first >> 7) | ((second >> 6) & 0x02);
+
+                        first <<= 1;
+                        second <<= 1;
+                    }
+                    index += 2
+                }
+            }
+
+            self.image_offset += 160 * 8;
         }
     }
 }